few refactor structure
